Move PathParamWidget into its own module
